Modifying get_rosetta.sh to reflect the new layout.
In light of the recent agreement to not use submodules, this script
has been updated to create a top-level directory and clone each repo
into it.  Internally, this represents a more substantial change as
the cloning and actual configuration has been moved into a for loop.
Additionally, some pretty ASCII art has been added for each repo to
make sure that running this script is "delightful".