Allowing ff-merges on master.
ff-merges make pulling work properly.  The rationale for disabling ff-merges was to prevent an edge case where someone might merge back to master and push before any other development happens, but the consequence is that every time you pull, you'll have to write a message.  If the edge case is a problem, we will address it then.